+++ client/components/RemoveUser.js
@@ -1,1 +1,7 @@
 //MAKE THIS ADMIN TRUE
+
+import React from 'react'
+
+const RemoveUser = () => <div />
+
+export default RemoveUser
