+++ client/components/index.js
@@ -9,3 +9,5 @@ export {Login, Signup} from './auth-form'
 export {default as AllTamagotchis} from './AllTamagotchis'
 export {default as SingleTamagotchi} from './SingleTamagotchi'
 export {default as CartOrder} from './CartOrder'
+export {default as AllUsers} from './AllUsers'
+export {default as SingleUser} from './SingleUser'
